# More efficient serialization of `BigDecimal` values