Before:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                        (size)   Mode  Cnt         Score         Error  Units
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc        1  thrpt    5  34399587.205 ± 3415434.195  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc       10  thrpt    5  22368590.579 ±  569069.367  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc      100  thrpt    5    697966.833 ±   38907.497  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc     1000  thrpt    5     31223.980 ±     224.406  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc    10000  thrpt    5      1254.372 ±     115.803  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc   100000  thrpt    5        47.623 ±       1.166  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc  1000000  thrpt    5         1.587 ±       0.056  ops/s
```
After:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                        (size)   Mode  Cnt         Score         Error  Units
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc        1  thrpt    5  45240545.379 ± 4583799.744  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc       10  thrpt    5  23181196.206 ± 2347618.368  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc      100  thrpt    5   1079613.505 ±    9818.976  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc     1000  thrpt    5     39964.378 ±    2360.273  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc    10000  thrpt    5      1360.489 ±       8.374  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc   100000  thrpt    5        49.240 ±       3.797  ops/s
[info] BigDecimalBenchmark.writeJsoniterScalaPrealloc  1000000  thrpt    5         1.654 ±       0.071  ops/s
```